+++ src/main/java/com/jedu_lima/EstoqueAPI/model/CriarProdutoCadastroEntradaDto.java
@@ -1,6 +1,8 @@
 package com.jedu_lima.EstoqueAPI.model;
 
-public record CriarProdutoCadastroEntradaDto(Long codigoDeBarras, String nome, Double valorCompra,
-		Integer quantidadeTotal, Integer porcentagemSobreVenda) {
+import java.math.BigDecimal;
+
+public record CriarProdutoCadastroEntradaDto(Long codigoDeBarras, String nome, BigDecimal valorCompra,
+		Integer quantidadeTotal, Double porcentagemSobreVenda) {
 
 }
