+++ src/main/java/com/jedu_lima/EstoqueAPI/model/CriarProdutoSaidaDto.java
@@ -1,10 +1,11 @@
 package com.jedu_lima.EstoqueAPI.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import com.jedu_lima.EstoqueAPI.entity.ProdutoCadastro;
 
-public record CriarProdutoSaidaDto(ProdutoCadastro produto, Integer quantidadeSaida, Double totalVenda,
+public record CriarProdutoSaidaDto(ProdutoCadastro produto, Integer quantidadeSaida, BigDecimal totalVenda,
 		LocalDate dataSaida) {
 
 }
